Day_8/Langchain/requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-google-genai
langchain-community
